fix evaluate table and pull out BuildRow into its own component
